move fallbacks to auth/httpfallback.go
